4.9. Подписки внутри view, рефакторинг утилит.
