+++ src/view/load-more-button-view.js
@@ -14,4 +14,14 @@ export default class LoadMoreButtonView extends AbstractView {
   get template() {
     return createLoadMoreButtonTemplate();
   }
+
+  setClickHandler = (callback) => {
+    this._callback.click = callback;
+    this.element.addEventListener('click', this.#clickHandler);
+  };
+
+  #clickHandler = (evt) => {
+    evt.preventDefault();
+    this._callback.click();
+  };
 }
